Extract advanced-settings library file
